# improvement: Use yaml.CLoader to load yaml files when available.