Hi @baberabb, this PR contains a speedup for the TaskManager `__init__` which is quite slow atm - it's quite annoying to wait 15+ seconds for an `lm_eval --task list`.

# Description

I briefly looked into it, turns out the longest time is spend parsing (read + graph construction) the (currently) `6625` yaml files under `tasks`. I suggest here a couple changes to switch from `yaml.FullLoader` to `yaml.CLoader` which as far as I can tell works the same - the only particular thing `lm_eval` needs are the custom constructors and those can be added to a `CLoader` as well. The one difference is that `CLoader` does not have a `.name` attribute (which points to the yaml file in the FullLoader) so I had to change the `import_function` accordingly by passing the `yaml_path` through.

# Runtimes

Comparison before and after the change

## Listing tasks [local]

Before (in `main`):

```bash
$ time lm_eval --task list
real    0m20.034s
user    0m19.182s
sys     0m2.114s
```

in this branch:

```bash
$ time lm_eval --task list
real    0m6.361s
user    0m6.984s
sys     0m1.514s
```

## Running Tests [local]

I haven't looked too much into it but I would expect a `TaskManager` is initialized more than once in testing. 

Before:
```bash
$ time python -m pytest --showlocals -s -vv -n=auto --ignore=tests/models/test_neuralmagic.py --ignore=tests/models/
============================================================================= 89 passed, 4 skipped, 41 warnings in 185.87s (0:03:05) =============================================================================
real    3m6.054s
user    16m29.785s
sys     1m24.957s
```

```bash
$ time python -m pytest --showlocals -s -vv -n=auto --ignore=tests/models/test_neuralmagic.py --ignore=tests/models/test_openvino.py
============================================================================= 89 passed, 4 skipped, 41 warnings in 116.94s (0:01:56) =============================================================================
real    1m57.136s
user    5m52.046s
sys     1m28.895s
```

## Running Tests, Github Actions

Decent speedup ->

Before:

[Random py3.9 job](https://github.com/EleutherAI/lm-evaluation-harness/actions/runs/13676912126/job/38239356496): `7m 57s`

Now:

[this PRs py3.9 job](https://github.com/EleutherAI/lm-evaluation-harness/actions/runs/13737129440/job/38422166960?pr=2777): `4m 20s`