@baberabb 2025-03-11
This is great! I meant to do some profiling, but never got around to it. Quite a speed up :)

As we keep adding more tasks and configs, I've been thinking about how to keep everything organized without the library getting unwieldy and keep the  startup time in check. It's a bit tricky since the library needs to parse all configs to handle the groups, tags and `include`s properly. Maybe we could look into some kind of tiered approach with core and extended tasks? I'm still exploring options, so I'm totally open to ideas if anyone has any suggestions!